New html templates (and associated locale strings) Some little modifications in the Service Layer. Some addition in the Model (for html views).

+++ src/main/java/fr/jodev/elite/services/SystemService.java
@@ -7,7 +7,7 @@ import fr.jodev.elite.entities.Station;
 
 public interface SystemService {
 	SolarSystem createSolarSystem(String name);
-	SolarSystem createSolarSystem(String name, Long x, Long y, Long z);
+	SolarSystem createSolarSystem(String name, Double x, Double y, Double z);
 	SolarSystem updateSystem(fr.jodev.elite.model.SolarSystem system);
 	void deleteSolarSystem(SolarSystem sys);
 	List<SolarSystem> getByName(String name);
